requirements.txt
coreschema==0.0.4
Django==4.0.4
django-cors-headers==3.12.0
django-environ==0.8.1
django-filter==22.1
django-jazzmin==2.5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
django-rest-swagger==2.1.1
drf-yasg==1.20.0
Pillow==9.1.1
psycopg2-binary==2.9.3
pytest==7.2.0
